+++ Cargo.toml
@@ -111,7 +111,7 @@ vlog = "0.1.4"
 mime = "0.3.13"
 farmhash = "1.1.5"
 failure = "0.1"
-tipb = { git = "https://github.com/pingcap/tipb.git" }
+tipb = { git = "https://github.com/wshwsh12/tipb.git" ,branch = "rename"}
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
 
 tikv_alloc = { path = "components/tikv_alloc", default-features = false }
