+++ components/tidb_query/Cargo.toml
@@ -39,7 +39,7 @@ safemem = { version = "0.3", default-features = false }
 flate2 = { version = "1.0", features = ["zlib"], default-features = false }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
-tipb = { git = "https://github.com/pingcap/tipb.git" }
+tipb = { git = "https://github.com/wshwsh12/tipb.git" ,branch = "rename"}
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
 
 tipb_helper = { path = "../tipb_helper" }
