+++ components/tidb_query/src/codec/batch/lazy_column_vec.rs
@@ -112,9 +112,9 @@ impl LazyBatchColumnVec {
         Ok(size)
     }
 
-    /// Returns maximum encoded size in arrow format.
+    /// Returns maximum encoded size in chunk format.
     // TODO: Move to other place.
-    pub fn maximum_encoded_size_arrow(
+    pub fn maximum_encoded_size_chunk(
         &self,
         logical_rows: impl AsRef<[usize]>,
         output_offsets: impl AsRef<[u32]>,
@@ -122,7 +122,7 @@ impl LazyBatchColumnVec {
         let logical_rows = logical_rows.as_ref();
         let mut size = 0;
         for offset in output_offsets.as_ref() {
-            size += self.columns[(*offset) as usize].maximum_encoded_size_arrow(logical_rows)?;
+            size += self.columns[(*offset) as usize].maximum_encoded_size_chunk(logical_rows)?;
         }
         Ok(size)
     }
@@ -148,8 +148,8 @@ impl LazyBatchColumnVec {
         Ok(())
     }
 
-    /// Encode into arrow format.
-    pub fn encode_arrow(
+    /// Encode into chunk format.
+    pub fn encode_chunk(
         &mut self,
         logical_rows: impl AsRef<[usize]>,
         output_offsets: impl AsRef<[u32]>,
